[JBEAP-18450][WFLY-12985] Move fault-tolerance from the ee feature pack and dists to the full one
This removes fault-tolerance from the 'legacy' non-Galleon feature packs. Legacy f-ps are only maintained to help Keycloak (which doesn't need MP) until it can move to Galleon.

This requires overrides of the Galleon domain config and the high level feature groups used in it in order to integrate the MP variant of domain-host-excludes.