[JBEAP-18745][WFLY-12952] Register default readiness probe for deployment
When a deployment does not provide any MicroProfile readiness probe,
add a default one (which returns UP as long as the deployment is deployed).

This allow to configure the overall server readiness to be DOWN when
the server is started and becomes UP as soon as a deployment is
deployed.

This default probe is disabled when the MicroProfile Health TCK is run.

JIRA: https://issues.redhat.com/browse/WFLY-12952